RAG/requirements.txt
google-genai
pypdf
nltk
tiktoken
chromadb
python-dotenv
tqdm
fastapi
uvicorn
scikit-learn
numpy
python-multipart
pydantic
pymupdf